# openapi.yaml
openapi: "3.0.3"
info:
  title: WebPeel API
  description: |
    Fast web fetcher for AI agents. Convert any website to LLM-ready markdown, 
    search the web, crawl sites, extract structured data, and more.
    
    **Base URL:** `https://api.webpeel.dev`
    
    **Authentication:** Optional. Pass API key via `Authorization: Bearer wp_...` header.
    Free tier available without authentication (125 requests/week).
    
    **Firecrawl Compatible:** WebPeel supports Firecrawl-compatible endpoints at the same paths.
    Switch from Firecrawl by changing your base URL.
  version: "0.8.1"
  contact:
    name: WebPeel Support
    email: support@webpeel.dev
    url: https://webpeel.dev
  license:
    name: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: https://api.webpeel.dev
    description: Production API
  - url: http://localhost:3000
    description: Local / Self-hosted
security:
  - bearerAuth: []
  - {}
paths:
...
  /v1/search:
    get:
      operationId: search
      summary: Search the web
      description: |
        Search the web and get results with snippets.
        Supports DuckDuckGo (default, free) and Brave Search (requires API key via BYOK).
      tags: [Search]
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Search query
        - name: provider
          in: query
          schema:
            type: string
            enum: [duckduckgo, brave]
            default: duckduckgo
          description: "Search provider (default: duckduckgo)"
        - name: searchApiKey
          in: query
          schema:
            type: string
          description: API key for search provider (alternative to x-search-api-key header)
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
          description: Max results
        - name: categories
          in: query
          schema:
            type: string
          description: Search categories
        - name: time
          in: query
          schema:
            type: string
            enum: [d, w, m, y]
          description: Time filter (day, week, month, year)
        - name: country
          in: query
          schema:
            type: string
          description: Country code for geo-targeted results
        - name: location
          in: query
          schema:
            type: string
          description: Location for geo-targeted results
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
...
  /v1/crawl:
    post:
      operationId: startCrawl
      summary: Start a crawl job
      description: Start an async crawl job. Returns a job ID for polling.
      tags: [Crawl]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:
                  type: string
                  description: Starting URL
                limit:
                  type: integer
                  default: 10
                  description: Max pages to crawl
                maxDepth:
                  type: integer
                  default: 3
                  description: Max crawl depth
                includePaths:
                  type: array
                  items:
                    type: string
                  description: URL path patterns to include
                excludePaths:
                  type: array
                  items:
                    type: string
                  description: URL path patterns to exclude
      responses:
        "200":
          description: Crawl job started
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  status:
                    type: string
  /v1/crawl/{id}:
    get:
      operationId: getCrawlStatus
      summary: Get crawl job status
      tags: [Crawl]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job status and results
    delete:
      operationId: cancelCrawl
      summary: Cancel a crawl job
      tags: [Crawl]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job cancelled
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API key (wp_...) or Firecrawl key (fc-...)
  schemas:
...
    SearchResponse:
      type: object
      properties:
        results:
          type: array
          items:
            type: object
            properties:
              title:
                type: string
              url:
                type: string
              snippet:
                type: string
...